Add files via upload
Next Hike Project 1 comprises of Merge datasets, identify unique values, drop unnecessary columns, check the dimensions of the dataset, check the datatype of the dataset, check datatype summary.
We have also treated missing values, validated the correctness of the data.
It includes Data Acquisition and Data Wrangling on dataset 1 and dataset 2, worked on central tendency.  
Concatenate datasheet, skewness and correlation of the data.